pyproject.toml
[tool.poetry]
name = "django-avito"
version = "0.1.0"
description = ""
authors = ["Maks Boldirev <boldirevmax10@gmail.com>"]
readme = "README.md"
packages = [{include = "django_avito"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
